requirements.txt
Flask
python-dotenv
google-genai
